requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
paho-mqtt==1.6.1
redis==5.0.1
celery==5.3.4
flower==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Data visualization and analysis
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
plotly==5.17.0

# Email notifications
smtplib2==0.2.1

# Hardware communication
pyserial==3.5
pymodbus==3.5.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0 